Added _DeleteWhereMetadataEquals procedure
Added a stored procedure that can be used to delete all rows from all
tables that have the specified @metadataId. This is useful when you want
to get rid of data that never should have been loaded in the first
place. Note that in CRT these rows will still be present in the
database, but have 0 reliability.